refactor: reorganize calendar tests by calendar file for better maintainability
- Broke up monolithic calendar-specific-alignment-issues.test.ts into separate files
- Created test/calendars/ directory structure for organized calendar testing
- Each test file now depends on only one calendar JSON file:
  - test/calendars/warhammer.test.ts (WFRP intercalary day issues)
  - test/calendars/dark-sun.test.ts (month start alignment)
  - test/calendars/forbidden-lands.test.ts (season alignment)
  - test/calendars/exandrian.test.ts (Critical Role canon)
  - test/calendars/cross-calendar-consistency.test.ts (consistency across calendars)

This prepares for future externally defined calendars where tests could be
included alongside JSON definitions. All tests passing (666/666).

Addresses PR review feedback requesting calendar test organization.
